requirements.txt
awscli==1.18.56
awswrangler==1.0.4
bcrypt==3.1.7
boto==2.49.0
boto3==1.12.42
botocore==1.15.42
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
diagrams==0.10.0
docutils==0.15.2
fabric==2.5.0
fsspec==0.7.2
future==0.18.2
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-bigquery-storage==0.8.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
Google2Pandas==0.1.1
googleapis-common-protos==1.51.0
graphviz==0.13.2
grpcio==1.28.1
httplib2==0.17.3
idna==2.9
invoke==1.4.1
Jinja2==2.11.2
jmespath==0.9.5
MarkupSafe==1.1.1
numpy==1.18.2
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.3
pandas-gbq==0.13.1
paramiko==2.7.1
prettytable==0.7.2
protobuf==3.11.3
psycopg2-binary==2.8.5
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAthena==1.10.4
pycparser==2.20
pydata-google-auth==0.3.0
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3fs==0.4.2
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.13
sqlalchemy-redshift==0.7.7
tenacity==6.1.0
uritemplate==3.0.1
urllib3==1.25.9
